Package restructure & Add save test
